Moves usePromise into utils
