+++ src/getMstGql.ts
@@ -1,5 +1,5 @@
 import { createHttpClient } from "mst-gql"
-import { useContext, useState } from "react"
+import { useContext } from "react"
 import { SERVER_URI } from "./config"
 import { RootStore, RootStoreType } from "./models"
 import { StoreContext } from "./models/reactUtils"
@@ -30,28 +30,3 @@ if (process.env.NODE_ENV === "development") {
 }
 
 export const useStore = (): RootStoreType => useContext(StoreContext)
-
-export type UsePromiseResponse<T> = {
-  loading: boolean
-  data: T | undefined
-  error: Error | undefined
-}
-
-export const usePromise = <T>(promise: PromiseLike<T>): UsePromiseResponse<T> => {
-  const [loading, setLoading] = useState(true)
-  const [data, setData] = useState<T | undefined>(undefined)
-  const [error, setError] = useState<Error | undefined>(undefined)
-
-  promise.then(
-    (data) => {
-      setData(data)
-      setLoading(false)
-    },
-    (err: Error) => {
-      setError(err)
-      setLoading(false)
-    },
-  )
-
-  return { loading, data, error }
-}
